Class "objectPad" objects and methods to store and retrieve information

Objects of class "objectPad" can be used as storage of information. There are methods for storage and retrieval.

Objects from the Class

Objects can be created by calls of the form new("objectPad", ...). Currently the ... arguments are ignored and an empty object is created.

Objects of class "objectPad" have no user-level slots. Values are stored and retrieved with the function pad, for the former purpose using it on the left-hand side of an assignment.

Slots

.Data:

Object of class "list", inherited from "list".

Extends

Class "list", from data part. Class "vector", by class "list", distance 2.

Methods

initialize

signature(.Object = "objectPad"): ...

pad

signature(x = "objectPad", item = "missing"): ...

pad

signature(x = "objectPad", item = "ANY"): ...

pad<-

signature(x = "objectPad"): ...

Author

Georgi Boshnakov

Note

The methods for this class are set to work directly on objects from classes that have slot pad of class objectPad. Classes may define additional methods for the function pad (it is generic) to adapt it to their needs.

See also

Examples

# create a pad and store some info.
pa <- new("objectPad")
pad(pa,"apple") <- "fruit"
pad(pa,"sweets") <- "sweets"
# pad(pa,"apple","dentist") <- "keeps away from her/him"
# pad(pa,"sweets","dentist") <- "helps meeting her/him"
pad(pa,"young Gauss") <- 100*(100+1)/2
pad(pa,"layman") <- sum(1:100)

pad(pa,"apple")
#> [1] "fruit"
pad(pa,"sweets")
#> [1] "sweets"
# pad(pa,"apple","dentist")
# pad(pa,"sweets","dentist")
pad(pa,"young Gauss") - pad(pa,"layman")
#> [1] 0

# create a class of vectors with a pad
setClass("vecWithPad",representation(x="numeric",pad="objectPad"))
a <- new("vecWithPad",x=1:10)
pad(a,"mean") <- mean(a@x)    # Compute and store the mean, and
pad(a,"mean")                 # retrieve it
#> [1] 5.5